+++ NEWS.md
@@ -6,7 +6,7 @@
 - scaling controls with scale instead of manually resulting in sligthtly different estimates. (but still same p-value and t-values)
 - removed ridge regression, t-test and correlation codes since they did not work
 - removed automatic removal of NAs in the topics predictions (this should be handled explicitly).
-
+- topicsTest `default` to linear_regression if not the variable only contains 0s and 1s; i.e., now different tests can be applied to different axes. 
 
 # topics 0.30.3
 - saving settings in `dtm` for downstream use in other functions.